Se refactorizaron otras cosas, pero el objetivo no se cumplió.
