+++ front/package.json
@@ -10,6 +10,7 @@
     "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
+    "@capacitor/core": "^2.4.7",
     "@quasar/extras": "^1.0.0",
     "axios": "^0.18.1",
     "core-js": "^3.6.5",
@@ -18,7 +19,8 @@
     "request": "^2.88.2",
     "request-promise-native": "^1.0.9",
     "vuex": "^3.6.2",
-    "wavesurfer.js": "^4.6.0"
+    "wavesurfer.js": "^4.6.0",
+    "xmlhttprequest": "^1.8.0"
   },
   "devDependencies": {
     "@quasar/app": "^2.0.0",
