+++ front/src/pages/Search.vue
@@ -5,23 +5,9 @@
       <div v-bind:key="result.id" v-for="result in searchResults">
         <SearchResultYoutube v-if="result.type === 'video'" :result="result" />
         <SearchResultSound v-else-if="result.type === 'sound'" :result="result" />
-        <q-btn round @click="agregarSound(result)" color="positive" icon="add" />
-          <q-btn round v-if="result.type === 'sound'" @click="downloadFile({name: result.sound_name, sound_file_url: result.sound_file_url, type: 'sound', url: result.sound_id})" color="positive" icon="download" />
-          <q-btn round v-else-if="result.type === 'video'" @click="downloadFile({name: result.title, sound_file_url: '.mp3', type: 'video', url: result.id})" color="positive" icon="download" />
         <q-separator></q-separator>
       </div>
     </div>
-    <q-dialog
-      v-model="dialogPlaylist"
-      transition-show="slide-up"
-      transition-hide="slide-down"
-    >
-      <q-card style="width: 800px; max-width: 80vw;" class="container">
-        <q-card-section>
-          <Playlist mode= 'adding' @addToPlaylist="addToPlaylist"></Playlist>
-        </q-card-section>
-      </q-card>
-    </q-dialog>
   </div>
 </template>
 
@@ -29,12 +15,10 @@
 import { functions } from '../functions.js'
 import SearchResultYoutube from '../components/SearchResultYoutube.vue'
 import SearchResultSound from '../components/SearchResultSound.vue'
-import Playlist from './Playlist.vue'
-import SoundPlaylistService from '../services/SoundPlaylistService'
 
 export default {
   mixins: [functions],
-  components: { SearchResultYoutube, SearchResultSound, Playlist },
+  components: { SearchResultYoutube, SearchResultSound },
   data () {
     return {
       token: localStorage.getItem('token'),
@@ -69,29 +53,6 @@ export default {
     })
   },
   methods: {
-    agregarSound (sound) {
-      this.dialogPlaylist = true
-      this.sound = sound
-    },
-    async addToPlaylist (playlist) {
-      try {
-        const data = {}
-        if (this.sound.type === 'video') {
-          data.playlist_id = playlist.playlist_id
-          data.youtube_id = this.sound.id
-          data.token = this.token
-        } else {
-          data.playlist_id = playlist.playlist_id
-          data.sound_id = this.sound.sound_id
-          data.token = this.token
-        }
-        const request = await SoundPlaylistService.add(data)
-        if (request.status >= 200 && request.status < 300) this.alert('positive', 'Canción agregada correctamente')
-      } catch (error) {
-        console.log(error)
-      }
-      this.dialogPlaylist = false
-    }
   }
 }
 </script>
